+++ age/src/lib.rs
@@ -83,7 +83,7 @@
 //! ## Passphrase-based encryption
 //!
 //! ```
-//! use secrecy::Secret;
+//! use age::secrecy::Secret;
 //! use std::io::{Read, Write};
 //!
 //! # fn run_main() -> Result<(), ()> {
@@ -138,6 +138,9 @@
 #![deny(broken_intra_doc_links)]
 #![deny(missing_docs)]
 
+// Re-export crates that are used in our public API.
+pub use age_core::secrecy;
+
 mod error;
 mod format;
 mod identity;
@@ -177,8 +180,10 @@ pub mod plugin;
 #[cfg_attr(docsrs, doc(cfg(feature = "ssh")))]
 pub mod ssh;
 
-use age_core::format::{FileKey, Stanza};
-use secrecy::SecretString;
+use age_core::{
+    format::{FileKey, Stanza},
+    secrecy::SecretString,
+};
 
 /// A private key or other value that can unwrap an opaque file key from a recipient
 /// stanza.